added all car routes to main carFunctions folder, and added explanations for code. also deleted unnecessary folders, aswell as unnecessary routes in index.js
