+++ seeds/carInventorySeeds.json
@@ -15,11 +15,11 @@
   {
     "make": "porsche",
     "model": "boxster",
-    "years": 2022,
+    "year": 2022,
     "price": 249990.0,
     "mileage": 2797,
     "fossil_fuel": true,
-    "automatic_transmission": false,
+    "automatic": false,
     "engine_cylinders": 6,
     "color": "white",
     "body_type": "convertible",
@@ -28,11 +28,11 @@
   {
     "make": "porsche",
     "model": "carrera",
-    "years": 2019,
+    "year": 2019,
     "price": 299990.0,
     "mileage": 9325,
     "fossil_fuel": true,
-    "automatic_transmission": false,
+    "automatic": false,
     "engine_cylinders": 6,
     "color": "yellow",
     "body_type": "coupe",
@@ -41,11 +41,11 @@
   {
     "make": "porsche",
     "model": "gt2",
-    "years": 2018,
+    "year": 2018,
     "price": 709990.0,
     "mileage": 5933,
     "fossil_fuel": true,
-    "automatic_transmission": false,
+    "automatic": false,
     "engine_cylinders": 6,
     "color": "black",
     "body_type": "coupe",
@@ -54,11 +54,11 @@
   {
     "make": "porsche",
     "model": "targa",
-    "years": 2015,
+    "year": 2015,
     "price": 239990.0,
     "mileage": 73176,
     "fossil_fuel": true,
-    "automatic_transmission": false,
+    "automatic": false,
     "engine_cylinders": 6,
     "color": "black",
     "body_type": "targa",
@@ -67,11 +67,11 @@
   {
     "make": "porsche",
     "model": "targa",
-    "years": 2006,
+    "year": 2006,
     "price": 129990.0,
     "mileage": 93970,
     "fossil_fuel": true,
-    "automatic_transmission": true,
+    "automatic": true,
     "engine_cylinders": 6,
     "color": "grey",
     "body_type": "targa",
@@ -80,11 +80,11 @@
   {
     "make": "porsche",
     "model": "cayenne",
-    "years": 2018,
+    "year": 2018,
     "price": 193990.0,
     "mileage": 48126,
     "fossil_fuel": true,
-    "automatic_transmission": true,
+    "automatic": true,
     "engine_cylinders": 8,
     "color": "grey",
     "body_type": "wagon",
@@ -94,11 +94,11 @@
   {
     "make": "porsche",
     "model": "clubsport",
-    "years": 2000,
+    "year": 2000,
     "price": 229990.0,
     "mileage": 72928,
     "fossil_fuel": true,
-    "automatic_transmission": false,
+    "automatic": false,
     "engine_cylinders": 6,
     "color": "silver",
     "body_type": "coupe",
@@ -107,11 +107,11 @@
   {
     "make": "aston martin",
     "model": "volante",
-    "years": 2011,
+    "year": 2011,
     "price": 239990.0,
     "mileage": 31251,
     "fossil_fuel": true,
-    "automatic_transmission": true,
+    "automatic": true,
     "engine_cylinders": 12,
     "color": "silver",
     "body_type": "convertible",
@@ -120,11 +120,11 @@
   {
     "make": "aston martin",
     "model": "db9",
-    "years": 2012,
+    "year": 2012,
     "price": 189990.0,
     "mileage": 29452,
     "fossil_fuel": true,
-    "automatic_transmission": true,
+    "automatic": true,
     "engine_cylinders": 12,
     "color": "grey",
     "body_type": "coupe",
@@ -133,11 +133,11 @@
   {
     "make": "aston martin",
     "model": "db11",
-    "years": 2017,
+    "year": 2017,
     "price": 294990.0,
     "mileage": 26880,
     "fossil_fuel": true,
-    "automatic_transmission": true,
+    "automatic": true,
     "engine_cylinders": 12,
     "color": "white",
     "body_type": "coupe",
@@ -146,11 +146,11 @@
   {
     "make": "bentley",
     "model": "arnage",
-    "years": 2003,
+    "year": 2003,
     "price": 109990.0,
     "mileage": 58286,
     "fossil_fuel": true,
-    "automatic_transmission": true,
+    "automatic": true,
     "engine_cylinders": 8,
     "color": "black",
     "body_type": "sedan",
@@ -159,11 +159,11 @@
   {
     "make": "bentley",
     "model": "bentayga",
-    "years": 2017,
+    "year": 2017,
     "price": 259990.0,
     "mileage": 35970,
     "fossil_fuel": true,
-    "automatic_transmission": true,
+    "automatic": true,
     "engine_cylinders": 8,
     "color": "white",
     "body_type": "wagon",
@@ -172,11 +172,11 @@
   {
     "make": "bentley",
     "model": "mulsanne",
-    "years": 2015,
+    "year": 2015,
     "price": 379990.0,
     "mileage": 12755,
     "fossil_fuel": true,
-    "automatic_transmission": true,
+    "automatic": true,
     "engine_cylinders": 8,
     "color": "onyx",
     "body_type": "sedan",
@@ -185,11 +185,11 @@
   {
     "make": "maserati",
     "model": "gransport",
-    "years": 2018,
+    "year": 2018,
     "price": 99990.0,
     "mileage": 17094,
     "fossil_fuel": true,
-    "automatic_transmission": true,
+    "automatic": true,
     "engine_cylinders": 6,
     "color": "white",
     "body_type": "sedan",
@@ -198,11 +198,11 @@
   {
     "make": "mclaren",
     "model": "675lt",
-    "years": 2015,
+    "year": 2015,
     "price": 1700000.0,
     "mileage": 3622,
     "fossil_fuel": true,
-    "automatic_transmission": false,
+    "automatic": false,
     "engine_cylinders": 8,
     "color": "grey",
     "body_type": "coupe",
@@ -211,11 +211,11 @@
   {
     "make": "mclaren",
     "model": "p1",
-    "years": 2014,
+    "year": 2014,
     "price": 2849990.0,
     "mileage": 1228,
     "fossil_fuel": true,
-    "automatic_transmission": true,
+    "automatic": true,
     "engine_cylinders": 8,
     "color": "yellow",
     "body_type": "coupe",
@@ -224,11 +224,11 @@
   {
     "make": "mclaren",
     "model": "620r",
-    "years": 2021,
+    "year": 2021,
     "price": 549990.0,
     "mileage": 10064,
     "fossil_fuel": true,
-    "automatic_transmission": false,
+    "automatic": false,
     "engine_cylinders": 8,
     "color": "silver",
     "body_type": "coupe",
@@ -237,11 +237,11 @@
   {
     "make": "mclaren",
     "model": "spider",
-    "years": 2019,
+    "year": 2019,
     "price": 369990.0,
     "mileage": 7973,
     "fossil_fuel": true,
-    "automatic_transmission": true,
+    "automatic": true,
     "engine_cylinders": 8,
     "color": "blue",
     "body_type": "convertible",
@@ -250,11 +250,11 @@
   {
     "make": "lamborghini",
     "model": "aventador",
-    "years": 2019,
+    "year": 2019,
     "price": 3500000.0,
     "mileage": 1513,
     "fossil_fuel": true,
-    "automatic_transmission": true,
+    "automatic": true,
     "engine_cylinders": 12,
     "color": "grigio telesto",
     "body_type": "coupe",
@@ -263,11 +263,11 @@
   {
     "make": "ferrari",
     "model": "fiorano",
-    "years": 2007,
+    "year": 2007,
     "price": 299990.0,
     "mileage": 32792,
     "fossil_fuel": true,
-    "automatic_transmission": false,
+    "automatic": false,
     "engine_cylinders": 12,
     "color": "rosso corse",
     "body_type": "coupe",
